+++ HISTORY.md
@@ -1,5 +1,7 @@
 # 0.5.0
 ## 2013-12-20
+* [breaking] Removed use of `options.emulateJSON` and `options.data` in `clientSync`
+* [breaking] Removed `addApiParams` from `syncer`
 * [breaking] Moved `AppView` to `client/` from `shared/`.
 * [breaking] Added support for AMD (RequireJS), which resulted in many methods
   becoming asynchronous instead of synchronous (`Fetcher`, `BaseView`,